moved split tunnel initialization and variable definitions
